% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dACG.R
\name{dACG}
\alias{dACG}
\title{Density of the ACG distribution}
\usage{
dACG(X, A)
}
\arguments{
\item{X}{matrix of observations (data matrix).}

\item{A}{parameter matrix.}
}
\value{
f ACG density evaluated at X.
}
\description{
Density of the ACG distribution
}
\examples{
N <- 100
q <- 3
b <- 1
lambda_n <- c(.588,.421,0)
A <- diag(lambda_n)
Psinv = diag(q) + (2/b) * A
Psi = solve(Psinv)
Y <- rACG(N,Psi)
f <- dACG(Y,Psi)
}
