% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rACG.R
\name{rACG}
\alias{rACG}
\title{Simulation of the Angular Central Gaussian (ACG) distribution}
\usage{
rACG(N, Psi)
}
\arguments{
\item{N:}{number of simulated random vectors.}

\item{Psi:}{(q x q) matrix of parameters of the angular central Gaussian distribution}
}
\value{
Y matrix of nreps simulated vectors of length q.
}
\description{
This function simulates the Angular Central Gaussian (ACG) distribution
}
\examples{
N <- 100
q <- 3
b <- 1
lambda_n <- c(.588,.421,0)
A <- diag(lambda_n)
Psinv = diag(q) + (2/b) * A
Psi = solve(Psinv)
Y <- rACG(N,Psi)
}
