% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBing.R
\name{rBing}
\alias{rBing}
\title{This function simulates the q-dimensional Bingham distribution by means of the acceptance-rejection algorithm
developed by Kent, J.T., Ganeiber, A.M., and Mardia, K.V. (2013). "A new method to simulate the
Bingham and related distributions in directional data analysis with applications."
http://arxiv.org/abs/1310.8110.}
\usage{
rBing(nreps, b, A, M)
}
\description{
This function simulates the q-dimensional Bingham distribution by means of the acceptance-rejection algorithm
developed by Kent, J.T., Ganeiber, A.M., and Mardia, K.V. (2013). "A new method to simulate the
Bingham and related distributions in directional data analysis with applications."
http://arxiv.org/abs/1310.8110.
}
